requirements.txt
lark-parser>=1.1.7
click>=8.0.0
colorama>=0.4.6